Make into_server return (SrpServer, Proof) instead
Keeping the server proof around in the struct is unnecessary since it is
never used again, and it provides a cleaner API to get it as a return
value instead of having to manually query it from something that is
otherwise only used for the session key and reconnection stuff.

This is a breaking change in the public API and bumps the minor version.