Make features able to be enabled simultaneously
The initial thought was that enable both features at the same time would
necessarily be a user error in the process of adding this crate to a
single other crate in an entire application.
In reality it might be possible that two crates in the same application
both depend on this crate but with different feature flags.
Cargo would then only compile one crate with all feature flags set,[1]
which would lead to an unnecessary compile error.

Since the only reason for not exclusively using the rug crate is that it
might fail to compile for some systems, defaulting to it in case both
feature flags are set should be fine.

[1]: https://old.reddit.com/r/rust/comments/sgegah/cargo_features_have_to_be_additive/