Remove hex_literal dependency
This completely removes the `hex_literal` dependency as well as gets rid
of 2 doctests.

The doctests were not necessary for consumers of the crate since they
don't need to know what the large safe prime is. People who are
interested can look in the source. Doctests are additionally slow since
they're compiled into separate binaries.